+++ pigit/const.py
@@ -6,7 +6,7 @@ import platform
 
 
 __project__ = "pigit"
-__version__ = "1.7.4"
+__version__ = "1.7.5"
 __url__ = "https://github.com/zlj-zz/pigit.git"
 __uri__ = __url__